+++ lib/src/chewie_player.dart
@@ -1,12 +1,17 @@
 import 'dart:async';
 
 import 'package:chewie/src/chewie_progress_colors.dart';
+import 'package:chewie/src/material/models/options_translation.dart';
+import 'package:chewie/src/models/subtitle_model.dart';
+import 'package:chewie/src/notifiers/player_notifier.dart';
 import 'package:chewie/src/player_with_controls.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
+import 'package:provider/provider.dart';
 import 'package:video_player/video_player.dart';
 import 'package:wakelock/wakelock.dart';
+
+import 'material/models/option_item.dart';
 
 typedef ChewieRoutePageBuilder = Widget Function(
   BuildContext context,
@@ -35,12 +40,16 @@ class Chewie extends StatefulWidget {
 }
 
 class ChewieState extends State<Chewie> {
-  bool _isFullScreen = false;
+  bool _isFullScreen = false; 
+
+  bool get isControllerFullScreen => widget.controller.isFullScreen;
+  late PlayerNotifier notifier;
 
   @override
   void initState() {
     super.initState();
     widget.controller.addListener(listener);
+    notifier = PlayerNotifier.init();
   }
 
   @override
@@ -55,14 +64,20 @@ class ChewieState extends State<Chewie> {
       widget.controller.addListener(listener);
     }
     super.didUpdateWidget(oldWidget);
+    if (_isFullScreen != isControllerFullScreen) {
+      widget.controller._isFullScreen = _isFullScreen;
+    }
   }
 
   Future<void> listener() async {
-    if (widget.controller.isFullScreen && !_isFullScreen) {
-      _isFullScreen = true;
+    if (isControllerFullScreen && !_isFullScreen) {
+      _isFullScreen = isControllerFullScreen;
       await _pushFullScreenWidget(context);
     } else if (_isFullScreen) {
-      Navigator.of(context, rootNavigator: true).pop();
+      Navigator.of(
+        context,
+        rootNavigator: widget.controller.useRootNavigator,
+      ).pop();
       _isFullScreen = false;
     }
   }
@@ -71,7 +86,10 @@ class ChewieState extends State<Chewie> {
   Widget build(BuildContext context) {
     return _ChewieControllerProvider(
       controller: widget.controller,
-      child: const PlayerWithControls(),
+      child: ChangeNotifierProvider<PlayerNotifier>.value(
+        value: notifier,
+        builder: (context, w) => const PlayerWithControls(),
+      ),
     );
   }
 
@@ -111,15 +129,26 @@ class ChewieState extends State<Chewie> {
   ) {
     final controllerProvider = _ChewieControllerProvider(
       controller: widget.controller,
-      child: const PlayerWithControls(),
+      child: ChangeNotifierProvider<PlayerNotifier>.value(
+        value: notifier,
+        builder: (context, w) => const PlayerWithControls(),
+      ),
     );
 
     if (widget.controller.routePageBuilder == null) {
       return _defaultRoutePageBuilder(
-          context, animation, secondaryAnimation, controllerProvider);
+        context,
+        animation,
+        secondaryAnimation,
+        controllerProvider,
+      );
     }
     return widget.controller.routePageBuilder!(
-        context, animation, secondaryAnimation, controllerProvider);
+      context,
+      animation,
+      secondaryAnimation,
+      controllerProvider,
+    );
   }
 
   Future<dynamic> _pushFullScreenWidget(BuildContext context) async {
@@ -133,7 +162,10 @@ class ChewieState extends State<Chewie> {
       Wakelock.enable();
     }
 
-    await Navigator.of(context, rootNavigator: true).push(route);
+    await Navigator.of(
+      context,
+      rootNavigator: widget.controller.useRootNavigator,
+    ).push(route);
     _isFullScreen = false;
     widget.controller.exitFullScreen();
 
@@ -141,30 +173,35 @@ class ChewieState extends State<Chewie> {
     // so we do not need to check Wakelock.isEnabled.
     Wakelock.disable();
 
-    SystemChrome.setEnabledSystemUIOverlays(
-        widget.controller.systemOverlaysAfterFullScreen);
+    SystemChrome.setEnabledSystemUIMode(
+      SystemUiMode.manual,
+      overlays: widget.controller.systemOverlaysAfterFullScreen,
+    );
     SystemChrome.setPreferredOrientations(
-        widget.controller.deviceOrientationsAfterFullScreen);
+      widget.controller.deviceOrientationsAfterFullScreen,
+    );
   }
 
   void onEnterFullScreen() {
     final videoWidth = widget.controller.videoPlayerController.value.size.width;
-    final videoHeight =
-        widget.controller.videoPlayerController.value.size.height;
+    final videoHeight = widget.controller.videoPlayerController.value.size.height;
 
-    if (widget.controller.systemOverlaysOnEnterFullScreen != null) {
-      /// Optional user preferred settings
-      SystemChrome.setEnabledSystemUIOverlays(
-          widget.controller.systemOverlaysOnEnterFullScreen!);
-    } else {
-      /// Default behavior
-      SystemChrome.setEnabledSystemUIOverlays([]);
-    }
+    SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual, overlays: []);
+
+    // if (widget.controller.systemOverlaysOnEnterFullScreen != null) {
+    //   /// Optional user preferred settings
+    //   SystemChrome.setEnabledSystemUIMode(
+    //     SystemUiMode.manual,
+    //     overlays: widget.controller.systemOverlaysOnEnterFullScreen,
+    //   );
+    // } else {
+    //   /// Default behavior
+    //   SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual, overlays: SystemUiOverlay.values);
+    // }
 
     if (widget.controller.deviceOrientationsOnEnterFullScreen != null) {
       /// Optional user preferred settings
-      SystemChrome.setPreferredOrientations(
-          widget.controller.deviceOrientationsOnEnterFullScreen!);
+      SystemChrome.setPreferredOrientations(widget.controller.deviceOrientationsOnEnterFullScreen!);
     } else {
       final isLandscapeVideo = videoWidth > videoHeight;
       final isPortraitVideo = videoWidth < videoHeight;
@@ -207,6 +244,7 @@ class ChewieState extends State<Chewie> {
 class ChewieController extends ChangeNotifier {
   ChewieController({
     required this.videoPlayerController,
+    this.optionsTranslation,
     this.aspectRatio,
     this.autoInitialize = false,
     this.autoPlay = false,
@@ -218,7 +256,12 @@ class ChewieController extends ChangeNotifier {
     this.placeholder,
     this.overlay,
     this.showControlsOnInitialize = true,
+    this.showOptions = true,
+    this.optionsBuilder,
+    this.additionalOptions,
     this.showControls = true,
+    this.subtitle,
+    this.subtitleBuilder,
     this.customControls,
     this.errorBuilder,
     this.allowedScreenSleep = true,
@@ -226,16 +269,128 @@ class ChewieController extends ChangeNotifier {
     this.allowFullScreen = true,
     this.allowMuting = true,
     this.allowPlaybackSpeedChanging = true,
+    this.useRootNavigator = true,
     this.playbackSpeeds = const [0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2],
     this.systemOverlaysOnEnterFullScreen,
     this.deviceOrientationsOnEnterFullScreen,
     this.systemOverlaysAfterFullScreen = SystemUiOverlay.values,
     this.deviceOrientationsAfterFullScreen = DeviceOrientation.values,
     this.routePageBuilder,
-  }) : assert(playbackSpeeds.every((speed) => speed > 0),
-            'The playbackSpeeds values must all be greater than 0') {
+  }) : assert(
+          playbackSpeeds.every((speed) => speed > 0),
+          'The playbackSpeeds values must all be greater than 0',
+        ) {
     _initialize();
   }
+
+  ChewieController copyWith({
+    VideoPlayerController? videoPlayerController,
+    OptionsTranslation? optionsTranslation,
+    double? aspectRatio,
+    bool? autoInitialize,
+    bool? autoPlay,
+    Duration? startAt,
+    bool? looping,
+    bool? fullScreenByDefault,
+    ChewieProgressColors? cupertinoProgressColors,
+    ChewieProgressColors? materialProgressColors,
+    Widget? placeholder,
+    Widget? overlay,
+    bool? showControlsOnInitialize,
+    bool? showOptions,
+    Future<void> Function(BuildContext, List<OptionItem>)? optionsBuilder,
+    List<OptionItem> Function(BuildContext)? additionalOptions,
+    bool? showControls,
+    Subtitles? subtitle,
+    Widget Function(BuildContext, dynamic)? subtitleBuilder,
+    Widget? customControls,
+    Widget Function(BuildContext, String)? errorBuilder,
+    bool? allowedScreenSleep,
+    bool? isLive,
+    bool? allowFullScreen,
+    bool? allowMuting,
+    bool? allowPlaybackSpeedChanging,
+    bool? useRootNavigator,
+    List<double>? playbackSpeeds,
+    List<SystemUiOverlay>? systemOverlaysOnEnterFullScreen,
+    List<DeviceOrientation>? deviceOrientationsOnEnterFullScreen,
+    List<SystemUiOverlay>? systemOverlaysAfterFullScreen,
+    List<DeviceOrientation>? deviceOrientationsAfterFullScreen,
+
+  Widget Function(
+      BuildContext,
+      Animation<double>,
+      Animation<double>,
+      _ChewieControllerProvider,
+    )?
+        routePageBuilder,
+  }) {
+    return ChewieController(
+      videoPlayerController: videoPlayerController ?? this.videoPlayerController,
+      optionsTranslation: optionsTranslation ?? this.optionsTranslation,
+      aspectRatio: aspectRatio ?? this.aspectRatio,
+      autoInitialize: autoInitialize ?? this.autoInitialize,
+      autoPlay: autoPlay ?? this.autoPlay,
+      startAt: startAt ?? this.startAt,
+      looping: looping ?? this.looping,
+      fullScreenByDefault: fullScreenByDefault ?? this.fullScreenByDefault,
+      cupertinoProgressColors: cupertinoProgressColors ?? this.cupertinoProgressColors,
+      materialProgressColors: materialProgressColors ?? this.materialProgressColors,
+      placeholder: placeholder ?? this.placeholder,
+      overlay: overlay ?? this.overlay,
+      showControlsOnInitialize: showControlsOnInitialize ?? this.showControlsOnInitialize,
+      showOptions: showOptions ?? this.showOptions,
+      optionsBuilder: optionsBuilder ?? this.optionsBuilder,
+      additionalOptions: additionalOptions ?? this.additionalOptions,
+      showControls: showControls ?? this.showControls,
+      subtitle: subtitle ?? this.subtitle,
+      subtitleBuilder: subtitleBuilder ?? this.subtitleBuilder,
+      customControls: customControls ?? this.customControls,
+      errorBuilder: errorBuilder ?? this.errorBuilder,
+      allowedScreenSleep: allowedScreenSleep ?? this.allowedScreenSleep,
+      isLive: isLive ?? this.isLive,
+      allowFullScreen: allowFullScreen ?? this.allowFullScreen,
+      allowMuting: allowMuting ?? this.allowMuting,
+      allowPlaybackSpeedChanging:
+          allowPlaybackSpeedChanging ?? this.allowPlaybackSpeedChanging,
+      useRootNavigator: useRootNavigator ?? this.useRootNavigator,
+      playbackSpeeds: playbackSpeeds ?? this.playbackSpeeds,
+      systemOverlaysOnEnterFullScreen: systemOverlaysOnEnterFullScreen ?? this.systemOverlaysOnEnterFullScreen,
+      deviceOrientationsOnEnterFullScreen: deviceOrientationsOnEnterFullScreen ?? this.deviceOrientationsOnEnterFullScreen,
+      systemOverlaysAfterFullScreen: systemOverlaysAfterFullScreen ?? this.systemOverlaysAfterFullScreen,
+      deviceOrientationsAfterFullScreen: deviceOrientationsAfterFullScreen ?? this.deviceOrientationsAfterFullScreen,
+      routePageBuilder: routePageBuilder ?? this.routePageBuilder,
+    );
+  }
+
+  /// If false, the options button in MaterialUI and MaterialDesktopUI
+  /// won't be shown.
+  final bool showOptions;
+
+  /// Pass your translations for the options like:
+  /// - PlaybackSpeed
+  /// - Subtitles
+  /// - Cancel
+  ///
+  /// Buttons
+  ///
+  /// These are required for the default `OptionItem`'s
+  final OptionsTranslation? optionsTranslation;
+
+  /// Build your own options with default chewieOptions shiped through
+  /// the builder method. Just add your own options to the Widget
+  /// you'll build. If you want to hide the chewieOptions, just leave them
+  /// out from your Widget.
+  final Future<void> Function(BuildContext context, List<OptionItem> chewieOptions)? optionsBuilder;
+
+  /// Add your own additional options on top of chewie options
+  final List<OptionItem> Function(BuildContext context)? additionalOptions;
+
+  /// Define here your own Widget on how your n'th subtitle will look like
+  final Widget Function(BuildContext context, dynamic subtitle)? subtitleBuilder;
+
+  /// Add a List of Subtitles here in `Subtitles.subtitle`
+  Subtitles? subtitle;
 
   /// The controller for the video you want to play
   final VideoPlayerController videoPlayerController;
@@ -264,8 +419,7 @@ class ChewieController extends ChangeNotifier {
 
   /// When the video playback runs into an error, you can build a custom
   /// error message.
-  final Widget Function(BuildContext context, String errorMessage)?
-      errorBuilder;
+  final Widget Function(BuildContext context, String errorMessage)? errorBuilder;
 
   /// The Aspect Ratio of the Video. Important to get the correct size of the
   /// video!
@@ -306,6 +460,9 @@ class ChewieController extends ChangeNotifier {
   /// Defines if the playback speed control should be shown
   final bool allowPlaybackSpeedChanging;
 
+  /// Defines if push/pop navigations use the rootNavigator
+  final bool useRootNavigator;
+
   /// Defines the set of allowed playback speeds user can change
   final List<double> playbackSpeeds;
 
@@ -325,8 +482,7 @@ class ChewieController extends ChangeNotifier {
   final ChewieRoutePageBuilder? routePageBuilder;
 
   static ChewieController of(BuildContext context) {
-    final chewieControllerProvider = context
-        .dependOnInheritedWidgetOfExactType<_ChewieControllerProvider>()!;
+    final chewieControllerProvider = context.dependOnInheritedWidgetOfExactType<_ChewieControllerProvider>()!;
 
     return chewieControllerProvider.controller;
   }
@@ -340,8 +496,7 @@ class ChewieController extends ChangeNotifier {
   Future _initialize() async {
     await videoPlayerController.setLooping(looping);
 
-    if ((autoInitialize || autoPlay) &&
-        !videoPlayerController.value.isInitialized) {
+    if ((autoInitialize || autoPlay) && !videoPlayerController.value.isInitialized) {
       await videoPlayerController.initialize();
     }
 
@@ -408,6 +563,10 @@ class ChewieController extends ChangeNotifier {
   Future<void> setVolume(double volume) async {
     await videoPlayerController.setVolume(volume);
   }
+
+  void setSubtitle(List<Subtitle> newSubtitle) {
+    subtitle = Subtitles(newSubtitle);
+  }
 }
 
 class _ChewieControllerProvider extends InheritedWidget {
@@ -420,6 +579,5 @@ class _ChewieControllerProvider extends InheritedWidget {
   final ChewieController controller;
 
   @override
-  bool updateShouldNotify(_ChewieControllerProvider old) =>
-      controller != old.controller;
+  bool updateShouldNotify(_ChewieControllerProvider old) => controller != old.controller;
 }