+++ lib/src/material/models/index.dart
@@ -1,0 +1,2 @@
+export 'option_item.dart';
+export 'options_translation.dart';